// src/types/serviceKeys.ts
export type ServiceKey =
  | "muhurt_namakaranam"
  | "muhurt_annaprasana"
  | "muhurt_kesa_khandana"
  | "muhurt_karna_vedha"
  | "muhurt_aksharabhyasa"
  | "muhurt_gruha_pravesh"
  | "muhurt_bhoomi_pooja"
  | "muhurt_vivaha"
  | "muhurt_other"
  | "numerology_complete"
  | "numerology_name_correction"
  | "numerology_baby_name"
  | "numerology_match_making"
  | "numerology_business"
  | "numerology_family"
  | "gems"
  | "horoscope"
  | "pooja";